AIRFLOW-4809 | s3_delete_objects_operator and empty keys
Current implementation of s3_delete_objects_operator fails
on empty list of keys. In case when list of keys to remove
comes from dynamic task (like s3_list_operator via XCom)
empty list of keys should be also considered valid.

Currently it throws an exception from boto3 (malformed XML).

The patch changes the behaviour so the operator exits
immedietely if there are no keys to remove.

+++ airflow/contrib/operators/s3_delete_objects_operator.py
@@ -77,6 +77,9 @@ class S3DeleteObjectsOperator(BaseOperator):
     def execute(self, context):
         s3_hook = S3Hook(aws_conn_id=self.aws_conn_id, verify=self.verify)
 
+        if isinstance(self.keys, list) and len(self.keys) == 0:
+            return
+
         response = s3_hook.delete_objects(bucket=self.bucket, keys=self.keys)
 
         deleted_keys = [x['Key'] for x in response.get("Deleted", [])]
